// src/app/model/user.ts
import { SharedDTO } from './sharedDTO';
export class User {
    username: string='';
    datasource: string='';
    token: string='';
    idUsuarioUra: number=0;
    webservice: string='';
    packageModel: string='';
    mostrarInfoArea: boolean=false;
    tokenCel: string='';
    metodo: string='';
    idAplica: number=0;
    aplicacion: string='';
    usuarioMesa: boolean=false;
    mail:string='';

}